refactor : 컨트롤러를 composite 패턴으로 변환.
