+++ src/main/java/webserver/RequestHandler.java
@@ -1,8 +1,9 @@
 package webserver;
 
+import logics.Controller;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import logics.Controller;
+import logics.RootController;
 import utils.requests.HttpRequest;
 import utils.requests.HttpRequestVersion1;
 import utils.response.HttpResponseVersion1;
@@ -34,8 +35,8 @@ public class RequestHandler implements Runnable {
     private void proceed(BufferedReader br, DataOutputStream dos) throws IOException{
         try{
             HttpRequest httpRequest = HttpRequestVersion1.readFrom(br);
-            Controller controller = new Controller();
-            controller.makeResponse(httpRequest).respond(dos);
+            Controller rootController = new RootController();
+            rootController.makeResponse(httpRequest).respond(dos);
         } catch(RuntimeException e){ // If RuntimeException Occurs,
             new HttpResponseVersion1().setResponseCode(400).respond(dos); // Respond 400(Bad Request)
         }
